Move trace processor to its own file, add ml_app arg to llmobs SDK
