+++ ddtrace/llmobs/_constants.py
@@ -2,6 +2,7 @@ SPAN_KIND = "_ml_obs.meta.span.kind"
 SESSION_ID = "_ml_obs.session_id"
 METRICS = "_ml_obs.metrics"
 TAGS = "_ml_obs.tags"
+ML_APP = "_ml_obs.meta.ml_app"
 
 MODEL_NAME = "_ml_obs.meta.model_name"
 MODEL_PROVIDER = "_ml_obs.meta.model_provider"
